Split huge NearestNeighborsQuerySpec into multiple Suites
