+++ elastiknn-testing/src/main/resources/logback-test.xml
@@ -1,9 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <configuration>
     <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <layout class="ch.qos.logback.classic.PatternLayout">
+        <encoder>
             <Pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} %msg%n</Pattern>
-        </layout>
+        </encoder>
     </appender>
     <logger name="com.sksamuel.elastic4s" level="WARN"/>
+    <logger name="ch.qos.logback.classic" level="WARN"/>
 </configuration>
