+++ elastiknn-testing/src/test/scala/com/klibisz/elastiknn/query/RecallSuite.scala
@@ -24,7 +24,7 @@ import scala.util.hashing.MurmurHash3.orderedHash
   *   the same query, I have seen different results at times. This seems to be an effect at the Elasticsearch level.
   *   I've tested at the Lucene (sans ES) level and that seems to be reliably deterministic.
   */
-class NearestNeighborsQueryRecallSuite extends AsyncFunSuite with Matchers with ElasticAsyncClient {
+class RecallSuite extends AsyncFunSuite with Matchers with ElasticAsyncClient {
 
   // Each test case consists of setting up one Mapping and then running several queries against that mapping.
   // Each query has an expected recall that will be checked.
